some renames, moves and reorderings
